move converter to its own file
